corrigindo ddl

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS endereco(
-    id INT AUTO_INCREMENT NOT NULL,
+    id BIGINT AUTO_INCREMENT NOT NULL,
     estado VARCHAR(30) NOT NULL,
     municipio VARCHAR(30) NOT NULL,
     bairro VARCHAR(30) NOT NULL,
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS endereco(
 );
 
 CREATE TABLE IF NOT EXISTS produto(
-    id INT AUTO_INCREMENT NOT NULL,
+    id BIGINT AUTO_INCREMENT NOT NULL,
     largura INT NOT NULL,
     altura INT NOT NULL,
     comprimento INT NOT NULL,
@@ -20,19 +20,19 @@ CREATE TABLE IF NOT EXISTS produto(
 );
 
 CREATE TABLE IF NOT EXISTS pagamento(
-    id INT AUTO_INCREMENT NOT NULL,
+    id BIGINT AUTO_INCREMENT NOT NULL,
     metodo ENUM('DEBITO', 'CREDITO', 'BOLETO', 'PIX'),
     PRIMARY KEY(id)
 );
 
 
 CREATE TABLE IF NOT EXISTS servico(
-    id INT AUTO_INCREMENT NOT NULL,
+    id BIGINT AUTO_INCREMENT NOT NULL,
     descricao VARCHAR(250) NOT NULL,
-    endereco_busca_id INT NOT NULL,
-    endereco_entrega_id INT NOT NULL,
-    produto_id INT NOT NULL,
-    pagamento_id INT NOT NULL,
+    endereco_busca_id BIGINT NOT NULL,
+    endereco_entrega_id BIGINT NOT NULL,
+    produto_id BIGINT NOT NULL,
+    pagamento_id BIGINT NOT NULL,
     status VARCHAR(20) NOT NULL,
     PRIMARY KEY(id),
     FOREIGN KEY(endereco_busca_id) REFERENCES endereco(id),
@@ -42,12 +42,12 @@ CREATE TABLE IF NOT EXISTS servico(
 );
 
 CREATE TABLE IF NOT EXISTS orcamento(
-   id INT AUTO_INCREMENT NOT NULL,
+   id BIGINT AUTO_INCREMENT NOT NULL,
    custo_embalagem  DOUBLE NOT NULL,
    taxa_desistencia  DOUBLE NOT NULL,
    valor  DOUBLE NOT NULL,
    prazo  INT NOT NULL,
-   servico_id INT NOT NULL,
+   servico_id BIGINT NOT NULL,
    PRIMARY KEY(id),
    FOREIGN KEY(servico_id) REFERENCES servico(id)
 );
